requirements.txt
click==8.1.3
Flask==2.2.2
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.1
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
pyserial==3.5
Werkzeug==2.2.2
